Moving organisation resolvers and types in their own files
