+++ packages/itmat-commons/src/graphql/index.ts
@@ -1,4 +1,5 @@
 import * as appUsers from './appUsers';
+import * as organisation from './organisation';
 import * as curation from './curation';
 import * as fields from './fields';
 import * as files from './files';
@@ -10,6 +11,7 @@ import * as study from './study';
 import * as user from './user';
 import * as subscription from './subscription';
 export * from './appUsers';
+export * from './organisation';
 export * from './curation';
 export * from './fields';
 export * from './files';
@@ -20,4 +22,4 @@ export * from './query';
 export * from './study';
 export * from './user';
 export * from './subscription';
-export const GQLRequests = { ...subscription, ...appUsers, ...curation, ...fields, ...files, ...log, ...permission, ...projects, ...query, ...study, ...user };
+export const GQLRequests = { ...subscription, ...appUsers, ...organisation, ...curation, ...fields, ...files, ...log, ...permission, ...projects, ...query, ...study, ...user };
