+++ packages/itmat-commons/src/models/index.ts
@@ -5,6 +5,7 @@ import * as Log from './log';
 import * as Query from './query';
 import * as Study from './study';
 import * as UserModels from './user';
+import * as Organisation from './organisation';
 import * as Data from './data';
 
 export * from './field';
@@ -14,5 +15,6 @@ export * from './log';
 export * from './query';
 export * from './study';
 export * from './user';
+export * from './organisation';
 export * from './data';
-export { File, JobModels, Log, UserModels, Study, Query, Field, Data };
+export { File, JobModels, Log, UserModels, Organisation, Study, Query, Field, Data };
