+++ packages/itmat-commons/src/models/organisation.ts
@@ -1,0 +1,12 @@
+export type OrganisationMetadata = {
+    siteIDMarker?: string;
+}
+
+export interface IOrganisation {
+    id: string;
+    name: string;
+    shortname?: string;
+    containOrg: string | null;
+    deleted: number | null;
+    metadata: OrganisationMetadata;
+}
